+++ src/GuzzleManager.php
@@ -112,7 +112,7 @@ class GuzzleManager implements Hermes
      */
     protected function resolve($name, array $options = [])
     {
-        $config = array_merge_recursive(
+        $config = array_merge_recursive_unique(
             $this->configurationFor($name), $options
         );
 
@@ -199,14 +199,14 @@ class GuzzleManager implements Hermes
     protected function createJsonDriver(array $config)
     {
         return new Client($this->makeClientOptions(
-            array_merge_recursive($config, [
+            array_merge_recursive_unique([
                 'options' => [
                     'response_handler' => JsonResponse::class,
                 ],
                 'interceptors' => [
                     Interceptors\ResponseHandler::class,
                 ],
-            ])
+            ], $config)
         ));
     }